Refactor user-service main into dependency module
